chore: rename ipns::dns to ::dnslink
